format code

+++ contracts/common/Helpers.sol
@@ -1,7 +1,5 @@
-// SPDX-License-Identifier: GPL-3.0
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.12;
-
-/* solhint-disable no-inline-assembly */
 
 /**
  * returned data from validateUserOp.
@@ -20,7 +18,7 @@ struct ValidationData {
 
 //extract sigFailed, validAfter, validUntil.
 // also convert zero validUntil to type(uint48).max
-function _parseValidationData(uint validationData) pure returns (ValidationData memory data) {
+function _parseValidationData(uint256 validationData) pure returns (ValidationData memory data) {
     address aggregator = address(uint160(validationData));
     uint48 validUntil = uint48(validationData >> 160);
     if (validUntil == 0) {
@@ -31,7 +29,10 @@ function _parseValidationData(uint validationData) pure returns (ValidationData 
 }
 
 // intersect account and paymaster ranges.
-function _intersectTimeRange(uint256 validationData, uint256 paymasterValidationData) pure returns (ValidationData memory) {
+function _intersectTimeRange(uint256 validationData, uint256 paymasterValidationData)
+    pure
+    returns (ValidationData memory)
+{
     ValidationData memory accountValidationData = _parseValidationData(validationData);
     ValidationData memory pmValidationData = _parseValidationData(paymasterValidationData);
     address aggregator = accountValidationData.aggregator;
@@ -78,4 +79,3 @@ function calldataKeccak(bytes calldata data) pure returns (bytes32 ret) {
         ret := keccak256(mem, len)
     }
 }
-
